Modulo_1-TratamientoDato/EjerciciosExtra/006/python/requirements.txt
Faker==16.0.0
pandas==1.5.0
mysql-connector-python
